+++ src-tauri/crates/psysonic-library/src/lib.rs
@@ -10,6 +10,7 @@
 pub(crate) mod bulk_ingest;
 pub mod advanced_search;
 pub mod album_compilation_filter;
+pub mod browse_support;
 mod advanced_search_mood;
 pub mod analysis_backfill;
 pub mod artist_lossless_browse;
